refactor: merge all modules into one module
